+++ react/components/Link.tsx
@@ -53,16 +53,30 @@ class Link extends Component<Props & RenderContextProps> {
     this.props.onClick(event)
 
     const options: NavigateOptions = {
+      fallbackToWindowLocation: false,
       page,
       params,
       query,
-      to,
       scrollOptions,
-      fallbackToWindowLocation: false,
+      to,
     }
     if (navigate(options)) {
       event.preventDefault()
     }
+  }
+
+  public getHref = ({to, page, pages, params, query}: any) => {
+    if (to) {
+      return to
+    }
+    if (page) {
+      const path = pathFromPageName(page, pages, params) 
+      const qs = query ? `?${query}` : ''
+      if (path) {
+        return path + qs
+      }
+    }
+    return '#'
   }
 
   public render() {
@@ -75,8 +89,7 @@ class Link extends Component<Props & RenderContextProps> {
       runtime: { pages },
       ...linkProps
     } = this.props
-    const href = to || (page && pathFromPageName(page, pages, params)) || '#'
-    return <a href={href} {...linkProps} onClick={this.handleClick} />
+    return <a href={this.getHref({ to, page, pages, params, query })} {...linkProps} onClick={this.handleClick} />
   }
 }